requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
Pillow==10.0.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0 
rest-framework-simplejwt==0.0.2
shortuuid==1.0.11
six==1.16.0 
python-dateutil==2.8.2